initial commit
This dataset considers various features that affect the quality of air in the atmosphere. Different machine learning algorithms are taken into part to make prediction on some features like the presence of various harmful and useful gases. All this data is visualized with the help of heatmap, correlation matrix and even many minute things are performed. Linear Regression, Random Forest Regressor and Neural Network are the algorithms on which model is trained to give results. Results of the entire algorithm are also pretty much good, which defines that algorithms are well suited for the task. Neural Network also defines the suitable loss that has extreme dominance on the performance, to judge the quality of the air.