initial commit
This dataset looks at the various gases like NO, NO2, CO, SO2, O3, etc, and then checks how all this data is correlated to each other, features should lack multicollinearity. After all, this dataset has many null values which are filled to make data relevant. Dataset is basically tells the air quality in the city Mumbai. Target feature also has many missing values which are also operated. There are many aspects that are needed to be considered in cities like Mumbai so as to prevent the degradation of air quality. Due to global warming various gases like O3 presence is increasing in the air which definitely decreases the air quality